main/requirements.txt
pyyaml==6.0.1
opencv-python==4.8.0.76
pandas==2.0.3
scikit-image==0.21.0
scikit-learn==1.3.0
pycocoevalcap == 1.2
dgl == 1.1.1
h5py == 3.9.0
torch_geometric == 2.3.1
simple_parsing == 0.1.4
tensorflow == 2.13.0
nltk == 3.8.1
wandb == 0.15.8ß